# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reasoning/assessments/{assessmentId}:
    get:
      operationId: getReasoningAssessment
      tags: [reasoning]
      summary: Get one assessment with its items (no scoring keys exposed)
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAssessment"
  /reasoning/assessments/{assessmentId}/start:
    post:
      operationId: startReasoningAttempt
      tags: [reasoning]
      summary: Start (or resume) an attempt on a diagnostic assessment
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartReasoningBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningAttemptState"
  /reasoning/assessments/{assessmentId}/submit:
    post:
      operationId: submitReasoningAttempt
      tags: [reasoning]
      summary: Submit responses; records a pass and returns written feedback
      parameters:
        - name: assessmentId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReasoningBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReasoningResult"
...
components:
  schemas:
...
    ReasoningItem:
      type: object
      required: [id, position, type, prompt]
      properties:
        id: { type: integer }
        position: { type: integer }
        type:
          type: string
          enum: [mcq, short_answer]
        prompt: { type: string }
        options:
          type: ["array", "null"]
          items: { type: string }
          description: For mcq items — the answer choices.
        allowNote:
          type: ["boolean", "null"]
          description: For hybrid-format mcq items — when true, an optional short written note accompanies the choice.
    ReasoningAssessment:
      type: object
      required: [id, instrument, phase, title, instructions, items]
      properties:
        id: { type: integer }
        instrument:
          type: string
          enum: [subject, reasoning]
        phase:
          type: string
          enum: [before, during1, during2, after]
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        instructions: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningItem"
    ReasoningAttemptState:
      type: object
      required: [id, assessmentId, status, startedAt, items]
      properties:
        id: { type: integer }
        assessmentId: { type: integer }
        status:
          type: string
          enum: [in_progress, submitted]
        format:
          type: ["string", "null"]
          enum: [mc, hybrid, written, null]
          description: The response format chosen for this attempt.
        length:
          type: ["string", "null"]
          enum: [short, medium, long, null]
          description: The length (question count) chosen for this attempt.
        needsFormat:
          type: ["boolean", "null"]
          description: >-
            When true, no attempt exists yet and the client must prompt the
            student to choose a format, then re-call start with that format.
            In this case id is 0 and items is empty.
        startedAt: { type: string, format: date-time }
        submittedAt: { type: ["string", "null"], format: date-time }
        passed: { type: ["boolean", "null"] }
        feedback: { type: ["string", "null"] }
        headline:
          type: ["string", "null"]
          description: For a submitted attempt being reviewed — the score headline.
        metrics:
          type: ["array", "null"]
          description: For a submitted attempt being reviewed — the score metrics.
          items:
            $ref: "#/components/schemas/ReasoningMetric"
        review:
          type: ["array", "null"]
          description: For a submitted attempt being reviewed — per-question review with the student's answer and the correct answer.
          items:
            $ref: "#/components/schemas/ReasoningReviewItem"
        items:
          type: array
          description: >-
            The exact items to present for THIS attempt. Every attempt — first
            take and every retake — returns freshly generated questions of the
            same kind (same instrument and phase), so no question is ever
            repeated.
          items:
            $ref: "#/components/schemas/ReasoningItem"
    ReasoningResponseInput:
      type: object
      required: [itemId]
      properties:
        itemId: { type: integer }
        selectedIndex:
          type: ["integer", "null"]
          description: mcq — chosen option index.
        text:
          type: ["string", "null"]
          description: short_answer — the student's brief written response.
        note:
          type: ["string", "null"]
          description: hybrid mcq — an optional short written justification for the choice.
...
    StartReasoningBody:
      type: object
      properties:
        retake:
          type: boolean
          description: >-
            When true, begin a fresh attempt even if a previous attempt was
            already submitted. An in-progress attempt is still resumed.
        format:
          type: string
          enum: [mc, hybrid, written]
          description: >-
            The response format for the attempt: mc = multiple-choice only,
            hybrid = multiple-choice plus a short written note, written = brief
            open written responses. Required when starting a brand-new attempt;
            ignored when resuming an existing one. When omitted and no attempt
            exists yet, the server responds with needsFormat=true so the client
            can prompt for a format.
        length:
          type: string
          enum: [short, medium, long]
          description: >-
            How many questions the attempt should contain: short = a few
            questions, medium = the standard set, long = a thorough set. Applied
            when starting a brand-new attempt; defaults to medium when omitted.
    SubmitReasoningBody:
      type: object
      required: [responses]
      properties:
        responses:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningResponseInput"
    ReasoningMetric:
      type: object
      required: [label, value]
      properties:
        label: { type: string }
        value: { type: string }
        detail: { type: ["string", "null"] }
    ReasoningReviewItem:
      type: object
      required: [itemId, type, prompt]
      properties:
        itemId: { type: integer }
        type:
          type: string
          enum: [mcq, short_answer]
        prompt: { type: string }
        options:
          type: ["array", "null"]
          items: { type: string }
          description: mcq — the answer choices shown.
        selectedIndex:
          type: ["integer", "null"]
          description: mcq — the option index the student chose.
        correctIndex:
          type: ["integer", "null"]
          description: mcq — the correct option index.
        isCorrect:
          type: ["boolean", "null"]
          description: mcq/short_answer — whether the student's answer was judged correct.
        note:
          type: ["string", "null"]
          description: hybrid mcq — the optional justification the student wrote.
        text:
          type: ["string", "null"]
          description: short_answer — the student's written response.
        referenceAnswer:
          type: ["string", "null"]
          description: short_answer — a model answer the response was graded against.
        verdict:
          type: ["string", "null"]
          enum: [correct, partial, incorrect, null]
          description: short_answer — the grader's verdict on the written response.
        rationale:
          type: ["string", "null"]
          description: short_answer — the grader's one-line rationale.
    ReasoningResult:
      type: object
      required: [attemptId, passed, feedback, headline, metrics, review]
      properties:
        attemptId: { type: integer }
        passed: { type: boolean }
        feedback: { type: string }
        headline: { type: string }
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/ReasoningMetric"
        review:
          type: array
          description: Per-question review — each item with the student's answer and the correct answer.
          items:
            $ref: "#/components/schemas/ReasoningReviewItem"
...