# lib/api-zod/src/generated/types/reasoningAttemptState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasoningAttemptStateFormat } from './reasoningAttemptStateFormat';
import type { ReasoningAttemptStateLength } from './reasoningAttemptStateLength';
import type { ReasoningAttemptStateStatus } from './reasoningAttemptStateStatus';
import type { ReasoningItem } from './reasoningItem';
import type { ReasoningMetric } from './reasoningMetric';
import type { ReasoningReviewItem } from './reasoningReviewItem';

export interface ReasoningAttemptState {
  id: number;
  assessmentId: number;
  status: ReasoningAttemptStateStatus;
  /**
     * The response format chosen for this attempt.
     * @nullable
     */
  format?: ReasoningAttemptStateFormat;
  /**
     * The length (question count) chosen for this attempt.
     * @nullable
     */
  length?: ReasoningAttemptStateLength;
  /**
     * When true, no attempt exists yet and the client must prompt the student to choose a format, then re-call start with that format. In this case id is 0 and items is empty.
     * @nullable
     */
  needsFormat?: boolean | null;
  startedAt: Date;
  /** @nullable */
  submittedAt?: Date | null;
  /** @nullable */
  passed?: boolean | null;
  /** @nullable */
  feedback?: string | null;
  /**
     * For a submitted attempt being reviewed — the score headline.
     * @nullable
     */
  headline?: string | null;
  /**
     * For a submitted attempt being reviewed — the score metrics.
     * @nullable
     */
  metrics?: ReasoningMetric[] | null;
  /**
     * For a submitted attempt being reviewed — per-question review with the student's answer and the correct answer.
     * @nullable
     */
  review?: ReasoningReviewItem[] | null;
  /** The exact items to present for THIS attempt. Every attempt — first take and every retake — returns freshly generated questions of the same kind (same instrument and phase), so no question is ever repeated. */
  items: ReasoningItem[];
}