# lib/api-zod/src/generated/types/reasoningItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasoningItemType } from './reasoningItemType';

export interface ReasoningItem {
  id: number;
  position: number;
  type: ReasoningItemType;
  prompt: string;
  /**
     * For mcq items — the answer choices.
     * @nullable
     */
  options?: string[] | null;
  /**
     * For hybrid-format mcq items — when true, an optional short written note accompanies the choice.
     * @nullable
     */
  allowNote?: boolean | null;
}