# lib/api-zod/src/generated/types/reasoningResponseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReasoningResponseInput {
  itemId: number;
  /**
     * mcq — chosen option index.
     * @nullable
     */
  selectedIndex?: number | null;
  /**
     * short_answer — the student's brief written response.
     * @nullable
     */
  text?: string | null;
  /**
     * hybrid mcq — an optional short written justification for the choice.
     * @nullable
     */
  note?: string | null;
}