# lib/api-zod/src/generated/types/reasoningReviewItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReasoningReviewItemType } from './reasoningReviewItemType';
import type { ReasoningReviewItemVerdict } from './reasoningReviewItemVerdict';

export interface ReasoningReviewItem {
  itemId: number;
  type: ReasoningReviewItemType;
  prompt: string;
  /**
     * mcq — the answer choices shown.
     * @nullable
     */
  options?: string[] | null;
  /**
     * mcq — the option index the student chose.
     * @nullable
     */
  selectedIndex?: number | null;
  /**
     * mcq — the correct option index.
     * @nullable
     */
  correctIndex?: number | null;
  /**
     * mcq/short_answer — whether the student's answer was judged correct.
     * @nullable
     */
  isCorrect?: boolean | null;
  /**
     * hybrid mcq — the optional justification the student wrote.
     * @nullable
     */
  note?: string | null;
  /**
     * short_answer — the student's written response.
     * @nullable
     */
  text?: string | null;
  /**
     * short_answer — a model answer the response was graded against.
     * @nullable
     */
  referenceAnswer?: string | null;
  /**
     * short_answer — the grader's verdict on the written response.
     * @nullable
     */
  verdict?: ReasoningReviewItemVerdict;
  /**
     * short_answer — the grader's one-line rationale.
     * @nullable
     */
  rationale?: string | null;
}